Изменил тип price на BigDecimal в бд shopping_store_products; getProductsPrice изменил на POST

+++ commerce/shopping-store/src/main/java/ru/yandex/practicum/dal/entity/Product.java
@@ -6,6 +6,7 @@ import ru.yandex.practicum.model.ProductCategory;
 import ru.yandex.practicum.model.ProductState;
 import ru.yandex.practicum.model.QuantityState;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -44,7 +45,6 @@ public class Product {
     @Column(name = "product_category")
     private ProductCategory productCategory;
 
-    //TODO: поменять на BigDecimal
     @Column(name = "price")
-    private Double price;
+    private BigDecimal price;
 }
